requirements.txt
kafka-python
confluent-kafka
datasketch
functools32